Web/requirements.txt
flask
cv2
google-cloud-storage
numpy
pickle
